backend/requirements.txt
flask
flask-cors
requests
python-dotenv
boto3 
google-generativeai
pypdf==4.1.0
python-dotenv==1.0.1